requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework==3.15.1
psycopg==3.1.18
python-dotenv==1.0.1
python-usernames==1.0.0
sqlparse==0.5.0
typing_extensions==4.11.0


Оформить код по flake8, сделать аннотацию типов, использовать линтеры, docstrings. (Для генерации docstrings функций используйте шаблон, автоматически создаваемый расширением autodocstring.)

Исправьте формы так, чтобы пользователь вводил время в формате HH:MM, а при вызове метода clean у нас значения start и end преобразовывались в int. Также учтите, что мы переопределили поле ввода start и end. Может возникнуть ошибка, что мы неверно ввели время.

Напишите ViewSets Для API, аналогично Forms.

Неавторизованные пользователи могут только регистрироваться.
Авторизованный пациент может просматривать свой профиль, записываться на прием к врачу, просматривать диагнозы и их статусы.
Авторизованный и is_active = True доктор может просматривать свой профиль, просматривать свои расписания, просматривать пациентов, записанных на прием, создавать и редактировать расписание, создавать и редактировать диагнозы.
Авторизованный UserType.ADMIN может просматривать свой профиль, просматривать всех докторов и пациентов, создавать и редактировать докторов и пациентов.
Авторизованный UserType.SUPERUSER может просматривать свой профиль, просматривать всех докторов и пациентов, создавать и редактировать докторов и пациентов, создавать и редактировать админов и суперюзеров.
Авторизованный и is_active = False доктор не имеет доступа к API.
UserType.ADMIN не может создавать/изменять/удалять UserType.SUPERUSER
UserType.ADMIN не может создавать/изменять/удалять UserType
UserType.SUPERUSER может создавать/изменять/удалять UserType.SUPERUSER
UserType.SUPERUSER может создавать/изменять/удалять UserType
UserType.SUPERUSER может создавать/изменять/удалять UserType.ADMIN
UserType.DOCTOR может изменять только своё расписание
UserType.PATIENT может изменять только свой профиль
UserType.PATIENT и UserType.DOCTOR не могут изменять UserType